Scripts/Spells: Migrate & update non-generic quest spell scripts to zone files (#31100)
(cherry picked from commit 53daa087aedcc0f68c935628f181256d89914ae2)
