+++ src/server/scripts/Northrend/zone_sholazar_basin.cpp
@@ -25,6 +25,7 @@
 #include "Player.h"
 #include "ScriptedEscortAI.h"
 #include "SpellAuras.h"
+#include "SpellInfo.h"
 #include "SpellScript.h"
 #include "Vehicle.h"
 #include "GameObject.h"
@@ -779,6 +780,89 @@ class spell_sholazar_flight_to_sholazar : public SpellScript
     }
 };
 
+/*######
+## Quest 12634: Some Make Lemonade, Some Make Liquor
+######*/
+
+enum SomeMakeLemonadeSomeMakeLiquor
+{
+    SPELL_BANANAS_FALL_TO_GROUND    = 51836,
+    SPELL_ORANGE_FALLS_TO_GROUND    = 51837,
+    SPELL_PAPAYA_FALLS_TO_GROUND    = 51839,
+    SPELL_TRIGGER_SPAWN_DWARF       = 52078
+};
+
+// This should be scripted in DB but our action lists aren't good enough to properly support chance
+// 51840 - Despawn Fruit Tosser
+class spell_sholazar_despawn_fruit_tosser : public SpellScript
+{
+    bool Validate(SpellInfo const* /*spellEntry*/) override
+    {
+        return ValidateSpellInfo(
+        {
+            SPELL_BANANAS_FALL_TO_GROUND,
+            SPELL_ORANGE_FALLS_TO_GROUND,
+            SPELL_PAPAYA_FALLS_TO_GROUND,
+            SPELL_TRIGGER_SPAWN_DWARF
+        });
+    }
+
+    void HandleDummy(SpellEffIndex /*effIndex*/)
+    {
+        uint32 spellId = SPELL_BANANAS_FALL_TO_GROUND;
+        switch (urand(0, 3))
+        {
+            case 1: spellId = SPELL_ORANGE_FALLS_TO_GROUND; break;
+            case 2: spellId = SPELL_PAPAYA_FALLS_TO_GROUND; break;
+        }
+        // sometimes, if you're lucky, you get a dwarf
+        if (roll_chance(5))
+            spellId = SPELL_TRIGGER_SPAWN_DWARF;
+
+        GetHitUnit()->CastSpell(GetCaster(), spellId);
+    }
+
+    void Register() override
+    {
+        OnEffectHitTarget += SpellEffectFn(spell_sholazar_despawn_fruit_tosser::HandleDummy, EFFECT_0, SPELL_EFFECT_DUMMY);
+    }
+};
+
+/*######
+## Quest 12805: Salvaging Life's Strength
+######*/
+
+enum SalvagingLifesStength
+{
+    NPC_SHARD_KILL_CREDIT       = 29303
+};
+
+// 54190 - Lifeblood Dummy
+class spell_sholazar_lifeblood_dummy : public SpellScript
+{
+    bool Validate(SpellInfo const* spellInfo) override
+    {
+        return ValidateSpellInfo({ uint32(spellInfo->GetEffect(EFFECT_0).CalcValue()) }) &&
+            sObjectMgr->GetCreatureTemplate(NPC_SHARD_KILL_CREDIT);
+    }
+
+    void HandleScript(SpellEffIndex /*effIndex*/)
+    {
+        if (GetHitUnit()->HasAura(uint32(GetEffectValue())))
+        {
+            if (Player* caster = GetCaster()->ToPlayer())
+                caster->KilledMonsterCredit(NPC_SHARD_KILL_CREDIT);
+
+            GetHitUnit()->RemoveAurasDueToSpell(uint32(GetEffectValueAsInt()));
+        }
+    }
+
+    void Register() override
+    {
+        OnEffectHitTarget += SpellEffectFn(spell_sholazar_lifeblood_dummy::HandleScript, EFFECT_0, SPELL_EFFECT_SCRIPT_EFFECT);
+    }
+};
+
 void AddSC_sholazar_basin()
 {
     RegisterCreatureAI(npc_engineer_helice);
@@ -794,4 +878,6 @@ void AddSC_sholazar_basin()
     RegisterSpellScript(spell_sholazar_song_of_cleansing);
     RegisterSpellScript(spell_sholazar_lightning_strike);
     RegisterSpellScript(spell_sholazar_flight_to_sholazar);
+    RegisterSpellScript(spell_sholazar_despawn_fruit_tosser);
+    RegisterSpellScript(spell_sholazar_lifeblood_dummy);
 }
