+++ src/server/scripts/Northrend/zone_zuldrak.cpp
@@ -999,7 +999,8 @@ class spell_zuldrak_summon_nass : public SpellScript
 
 enum ConvocationAtZolHeb
 {
-    SPELL_AKALIS_STUN     = 52989
+    SPELL_AKALIS_STUN           = 52989,
+    SPELL_FLICKERING_FLAMES     = 53504
 };
 
 // 53010 - Convocation at Zol'Heb: Removef Akali's Stun
@@ -1018,6 +1019,138 @@ class spell_zuldrak_remove_akalis_stun : public SpellScript
     void Register() override
     {
         OnEffectHitTarget += SpellEffectFn(spell_zuldrak_remove_akalis_stun::HandleScript, EFFECT_0, SPELL_EFFECT_SCRIPT_EFFECT);
+    }
+};
+
+// 53350 - Quenching Mist
+class spell_zuldrak_quenching_mist : public AuraScript
+{
+    bool Validate(SpellInfo const* /*spellInfo*/) override
+    {
+        return ValidateSpellInfo({ SPELL_FLICKERING_FLAMES });
+    }
+
+    void HandleEffectPeriodic(AuraEffect const* /*aurEff*/)
+    {
+        GetTarget()->RemoveAurasDueToSpell(SPELL_FLICKERING_FLAMES);
+    }
+
+    void Register() override
+    {
+        OnEffectPeriodic += AuraEffectPeriodicFn(spell_zuldrak_quenching_mist::HandleEffectPeriodic, EFFECT_0, SPELL_AURA_PERIODIC_HEAL);
+    }
+};
+
+/*######
+## Quest 12661: Infiltrating Voltarus / 12669: So Far, So Bad / 12676: Sabotage / 12677: Hazardous Materials / 12713: Betrayal
+######*/
+
+// 55368 - Summon Stefan
+class spell_zuldrak_summon_stefan : public SpellScript
+{
+    void SetDest(SpellDestination& dest)
+    {
+        // Adjust effect summon position (sniff)
+        Position const offset = { 0.0f, 0.0f, 30.0f, 0.0f };
+        dest.RelocateOffset(offset);
+    }
+
+    void Register() override
+    {
+        OnDestinationTargetSelect += SpellDestinationTargetSelectFn(spell_zuldrak_summon_stefan::SetDest, EFFECT_0, TARGET_DEST_CASTER_BACK);
+    }
+};
+
+/*######
+## Quest 12527: Gluttonous Lurkers
+######*/
+
+enum GluttonousLurkers
+{
+    SPELL_SUMMON_GORGED_LURKING_BASILISK    = 50928
+};
+
+// 50894 - Zul'Drak Rat
+class spell_zuldrak_zuldrak_rat : public SpellScript
+{
+    bool Validate(SpellInfo const* /*spell*/) override
+    {
+        return ValidateSpellInfo({ SPELL_SUMMON_GORGED_LURKING_BASILISK });
+    }
+
+    void HandleScriptEffect(SpellEffIndex /* effIndex */)
+    {
+        if (GetHitAura() && GetHitAura()->GetStackAmount() >= GetSpellInfo()->StackAmount)
+        {
+            GetHitUnit()->CastSpell((Unit*) nullptr, SPELL_SUMMON_GORGED_LURKING_BASILISK, true);
+            if (Creature* basilisk = GetHitUnit()->ToCreature())
+                basilisk->DespawnOrUnsummon();
+        }
+    }
+
+    void Register() override
+    {
+        OnEffectHitTarget += SpellEffectFn(spell_zuldrak_zuldrak_rat::HandleScriptEffect, EFFECT_1, SPELL_EFFECT_SCRIPT_EFFECT);
+    }
+};
+
+/*######
+## Quest 12919: The Storm King's Vengeance
+######*/
+
+enum TheStormKingsVengeance
+{
+    SPELL_RIDE_GYMER            = 43671,
+    SPELL_GRABBED               = 55424,
+    SPELL_VARGUL_EXPLOSION      = 55569
+};
+
+// 55516 - Gymer's Grab
+class spell_zuldrak_gymers_grab : public SpellScript
+{
+    bool Validate(SpellInfo const* /*spell*/) override
+    {
+        return ValidateSpellInfo({ SPELL_RIDE_GYMER, SPELL_GRABBED });
+    }
+
+    void HandleScript(SpellEffIndex /*effIndex*/)
+    {
+        if (!GetHitCreature())
+            return;
+        CastSpellExtraArgs args(TRIGGERED_FULL_MASK);
+        args.AddSpellBP0(2);
+        GetHitCreature()->CastSpell(GetCaster(), SPELL_RIDE_GYMER, args);
+        GetHitCreature()->CastSpell(GetHitCreature(), SPELL_GRABBED, true);
+    }
+
+    void Register() override
+    {
+        OnEffectHitTarget += SpellEffectFn(spell_zuldrak_gymers_grab::HandleScript, EFFECT_0,  SPELL_EFFECT_SCRIPT_EFFECT);
+    }
+};
+
+// 55421 - Gymer's Throw
+class spell_zuldrak_gymers_throw : public SpellScript
+{
+    bool Validate(SpellInfo const* /*spell*/) override
+    {
+        return ValidateSpellInfo({ SPELL_VARGUL_EXPLOSION });
+    }
+
+    void HandleScript(SpellEffIndex /*effIndex*/)
+    {
+        Unit* caster = GetCaster();
+        if (caster->IsVehicle())
+            if (Unit* passenger = caster->GetVehicleKit()->GetPassenger(1))
+            {
+                passenger->ExitVehicle();
+                caster->CastSpell(passenger, SPELL_VARGUL_EXPLOSION, true);
+            }
+    }
+
+    void Register() override
+    {
+        OnEffectHitTarget += SpellEffectFn(spell_zuldrak_gymers_throw::HandleScript, EFFECT_0, SPELL_EFFECT_SCRIPT_EFFECT);
     }
 };
 
@@ -1044,4 +1177,9 @@ void AddSC_zuldrak()
     RegisterSpellScript(spell_zuldrak_breaking_through_summon_zulaman_gods_master);
     RegisterSpellScript(spell_zuldrak_summon_nass);
     RegisterSpellScript(spell_zuldrak_remove_akalis_stun);
+    RegisterSpellScript(spell_zuldrak_quenching_mist);
+    RegisterSpellScript(spell_zuldrak_summon_stefan);
+    RegisterSpellScript(spell_zuldrak_zuldrak_rat);
+    RegisterSpellScript(spell_zuldrak_gymers_grab);
+    RegisterSpellScript(spell_zuldrak_gymers_throw);
 }
